Calculate shapefile of Munich based on geojson containing districts
The city of Munich only provides a GeoJSON encoding it's districts. Since I need the city as a whole as shapefile for the USGS Earth Explorer, I calculate a shapefile based on the districts' borders. There is also a downsampling performed because the USGS Earth Explorer only allows shapefiles with <500 vertices.